Move UIImage extension to separate file
